Fix rubocop
