+++ app/controllers/spaces_controller.rb
@@ -15,7 +15,7 @@ class SpacesController < AuthenticateController # rubocop:disable Metrics/ClassL
     @space = Space.new
   end
 
-  def create
+  def create # rubocop:disable Metrics/AbcSize
     address_params = get_address_params(params)
     return redirect_to spaces_path, alert: t("address_search.didnt_find") if address_params.nil?
