+++ app/models/space.rb
@@ -35,7 +35,7 @@ class Space < ApplicationRecord # rubocop:disable Metrics/ClassLength
 
   validates :star_rating, numericality: { greater_than: 0, less_than: 6 }, allow_nil: true
   validates :url, url: { allow_blank: true, public_suffix: true }
-  validates_presence_of :title, :address, :post_address, :post_number, :lat, :lng, presence: true
+  validates :title, :address, :post_address, :post_number, :lat, :lng, presence: { presence: true }
 
   after_create do
     aggregate_facility_reviews
