+++ spec/support/tom_select.rb
@@ -1,5 +1,6 @@
-module TomSelect
+# frozen_string_literal: true
 
+module TomSelect
   # A helper for Capyabara tests that need to set values from a tom-select.js input.
   #
   # This is a really hacky approach using execute_javascript, but it works. Not sure if there's
@@ -12,7 +13,7 @@ module TomSelect
   # @example tom_select("#select_id", option_id: "2")
   # @example tom_select("#select_id", option_id: ["2", "10"]) # `multiple` input.
   def tom_select(select_selector, option_id:)
-    js_str = %Q(document.querySelector("#{select_selector}").tomselect.setValue(#{option_id.inspect}))
+    js_str = %(document.querySelector("#{select_selector}").tomselect.setValue(#{option_id.inspect}))
     execute_script(js_str)
   end
 end
